Move sample data into tests and flatten ExampleData
