+++ src/ExampleLib/Domain/ISummarisationPlanStore.cs
