+++ src/ExampleLib/Domain/SummarisationPlan.cs
