+++ src/ExampleLib/Domain/ThresholdType.cs
