+++ src/ExampleLib/Domain/ValidationRule.cs
@@ -1,3 +1,3 @@
-namespace ExampleData;
+namespace ExampleLib.Domain;
 
 public record ValidationRule(ValidationStrategy Strategy, double Threshold);
