+++ tests/ExampleLib.BDDTests/ValidationPlanFactorySteps.cs
@@ -1,4 +1,5 @@
 using ExampleData;
+using ExampleLib.Domain;
 using Microsoft.EntityFrameworkCore;
 using Reqnroll;
 using System.Collections.Generic;
