+++ tests/ExampleLib.Tests/ExampleData/Entities.cs
@@ -1,0 +1,19 @@
+namespace ExampleData;
+
+using ExampleLib.Domain;
+
+public class YourEntity : BaseEntity
+{
+    public string Name { get; set; } = string.Empty;
+    public DateTime Timestamp { get; set; }
+}
+
+public class Nanny
+{
+    public int Id { get; set; }
+    public string ProgramName { get; set; } = string.Empty;
+    public string Entity { get; set; } = string.Empty;
+    public double SummarizedValue { get; set; }
+    public DateTime DateTime { get; set; }
+    public Guid RuntimeID { get; set; }
+}
