+++ tests/ExampleLib.Tests/ExampleData/YourEntityConfiguration.cs
